backend/requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
google-generativeai
openai
pydantic
python-dotenv
